Removed some debugging stuff. Works more or less. Crashes randomly -> need to figure out what gives...

+++ src/videoReader/VideoReader.cpp
@@ -457,5 +457,5 @@ int VideoReader::getNumberOfFrames(){
 }
 
 int VideoReader::getNumberOfIndices(){
-	return frameIndices.size();
+	return (int) frameIndices.size();
 }
